Created localhostURL for the new deployment link, moved api from components folder into src, and fixed all broken imports
