+++ .env
@@ -1,0 +1,1 @@
+DATABASE_URL="mongodb+srv://administrator:c0Unt3r_strIKE666KEKW@cluster0.vbzif3g.mongodb.net/where_is_everyone?retryWrites=true&w=majority&appName=Cluster0"
