+++ src/utils/localhostURL.js
@@ -1,0 +1,3 @@
+// export const localhostURL = 'http://localhost:5000'
+
+export const localhostURL = "https://where-is-everyone-backend.vercel.app";
